Refactor single file into rakefile based rubygem
