+++ .gitignore
@@ -2,3 +2,14 @@
 items
 !items/.gitkeep
 items.json
+
+# RubyGem
+/.bundle/
+/.yardoc
+/Gemfile.lock
+/_yardoc/
+/coverage/
+/doc/
+/pkg/
+/spec/reports/
+/tmp/
