+++ cloudapp-export.gemspec
@@ -1,0 +1,29 @@
+# coding: utf-8
+lib = File.expand_path("../lib", __FILE__)
+$LOAD_PATH.unshift(lib) unless $LOAD_PATH.include?(lib)
+require "cloudapp-export/version"
+
+Gem::Specification.new do |spec|
+  spec.name          = "cloudapp-export"
+  spec.version       = CloudappExport::VERSION
+  spec.authors       = ["Marc Qualie"]
+  spec.email         = ["marc@marcqualie.com"]
+
+  spec.summary       = "A quick way to export all the data you have stored in CloudApp"
+  spec.homepage      = "https://github.com/marcqualie/cloudapp-export"
+  spec.license       = "MIT"
+
+  spec.files         = `git ls-files -z`.split("\x0").reject do |f|
+    f.match(%r{^(test|spec|features)/})
+  end
+  spec.bindir        = "exe"
+  spec.executables   = spec.files.grep(%r{^exe/}) { |f| File.basename(f) }
+  spec.require_paths = ["lib"]
+
+  spec.add_dependency "dotenv"
+  spec.add_dependency "net-http-digest_auth"
+
+  spec.add_development_dependency "bundler", "~> 1.15"
+  spec.add_development_dependency "rake", "~> 10.0"
+  spec.add_development_dependency "minitest", "~> 5.0"
+end
