+++ lib/cloudapp-export.rb
@@ -1,0 +1,8 @@
+require "cloudapp-export/request"
+require "cloudapp-export/response"
+require "cloudapp-export/version"
+
+module Cloudapp
+  module Export
+  end
+end
