+++ lib/cloudapp-export/version.rb
@@ -1,0 +1,3 @@
+module CloudappExport
+  VERSION = "0.1.0"
+end
