+++ test/cloudapp/export_test.rb
@@ -1,0 +1,11 @@
+require "test_helper"
+
+class CloudappExportTest < Minitest::Test
+  def test_that_it_has_a_version_number
+    refute_nil ::CloudappExport::VERSION
+  end
+
+  def test_it_does_something_useful
+    assert false
+  end
+end
