+++ test/test_helper.rb
@@ -1,0 +1,4 @@
+$LOAD_PATH.unshift File.expand_path("../../lib", __FILE__)
+require "cloudapp-export"
+
+require "minitest/autorun"
